packages/career-consultant/requirements-dev.txt
# Development and Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Required dependencies for the project
pyyaml==6.0.1
pandas==2.1.3
openpyxl==3.1.2

# Optional: Property-based testing
hypothesis==6.92.1
